Function to get original teacher and room
With the old version, it is impossible to get the original teacher & room if there is a change. The new functions "get_original_teacher" and "get_original_room" will return the original teacher / room if available, or the current teacher / room if the original one isn't available. 

I'm not quite sure if such a feature already existed, but I didn't find it and so I wrote my own code for that.